LuaHook-NG Initial commit
Initial commit of LuaHook-NG. This solution consists of:

1) HookIPHLPAPI - a trivial stub wrapper that loads the main library.

2) Lua5.1 - a static library of C++ified Lua5.1.1

3) LuaHook-NG - the library implementing the main functionality

Callback functionality is implemented but not (yet) functional and the
code is thus purely in a PoC state until fleshed out. The main goal
however of providing a generic framework that is game-agnostic is done.

It might be worth separating out the game-specific logic into a separate
library but since we're only supporting Payday2 for the moment, there's
no real point.

Microsoft Detours is an external dependency and is not included in the
project for license reasons. Design-wise, it could be easily
substituted.