Testsuite implemented, code fixes, API additions.
The test suite is now running using Google Test although it is still not
a comprehensive test suite, it tests many of the more error-prone
functions (especially those that require reimplementation or wrapping)

The test suite runs as a client library and so can be simply be dropped
in for doing a test run. Using this library helped to identify several
unknown issues in the implementation.

The next step will be to separate the Console functionality into a
client library