Miscellaneous code refactoring and minor fixes.
the LuaHookNG callback registration struct now has non-const members but
the struct is taken as a const. Callers cannot change their callbacks by
changing the struct after instantiating a LuaGame.

Modified DeveloperTools for the new call style and added checking for
nullptr returns when attempting to retrieve output from the Lua state.
Also refactored a few classes into their own files.

LuaHookNG now offers a raw C Closure registration for functions that do
not need to be passed the LuaInterface reference. this is ONLY for
optimising cases where a closure is to be invoked with the LuaInterface
passed as a (light)userdata to the closure.