Move Lua5.1 code into Git Submodule.
The Lua 5.1 code is essentially unchanging and not particularly suitable
to being included in the main codebase since it also happens to be game
specific. Consequently, it's now moved into a submodule.

I may do the same with other parts of the project, but for now there's
no real impetus to do so since the code isn't particularly large yet,
unlike Lua.